orm: move ValidateSequence to sequence.go
