+++ orm/sequence_test.go
@@ -3,7 +3,9 @@ package orm
 import (
 	"testing"
 
+	"github.com/iov-one/weave/errors"
 	"github.com/iov-one/weave/store"
+	"github.com/iov-one/weave/weavetest"
 	"github.com/iov-one/weave/weavetest/assert"
 )
 
@@ -43,4 +45,40 @@ func TestSequenceKeyFormat(t *testing.T) {
 		t.Fatal("sequence not found in store, invalid key?")
 	}
 
+}
+
+func TestValidateSequence(t *testing.T) {
+	cases := map[string]struct {
+		bytes   []byte
+		wantErr *errors.Error
+	}{
+		"success": {
+			bytes:   []byte{0, 1, 2, 3, 4, 5, 6, 7},
+			wantErr: nil,
+		},
+		"success with sequence": {
+			bytes:   weavetest.SequenceID(12345),
+			wantErr: nil,
+		},
+		"failure missing": {
+			bytes:   nil,
+			wantErr: errors.ErrEmpty,
+		},
+		"failure invalid length": {
+			bytes:   []byte{0, 1},
+			wantErr: errors.ErrInput,
+		},
+	}
+
+	for testName, tc := range cases {
+		t.Run(testName, func(t *testing.T) {
+			err := ValidateSequence(tc.bytes)
+			if !tc.wantErr.Is(err) {
+				t.Fatalf("unexpected error: %+v", err)
+			}
+			if tc.wantErr != nil {
+				return
+			}
+		})
+	}
 }
